# [serve] Use `MessageQueue` for yielding generator results in the replica