--- a/python/ray/serve/_private/http_util.py
+++ b/python/ray/serve/_private/http_util.py
@@ -131,10 +131,17 @@ class ASGIMessageQueue(Send):
     `get_messages_nowait` and `wait_for_message` is undefined behavior).
     """
 
-    def __init__(self):
+    def __init__(
+        self, *, write_thread_safe: bool = False, loop: asyncio.AbstractEventLoop = None
+    ):
         self._message_queue = asyncio.Queue()
         self._new_message_event = asyncio.Event()
         self._closed = False
+        self._write_thread_safe = write_thread_safe
+        self._loop = loop
+
+        if self._write_thread_safe and self._loop is None:
+            raise ValueError("If `write_thread_safe=True`, `loop` must be provided.")
 
     def close(self):
         """Close the queue, rejecting new messages.
@@ -146,16 +153,27 @@ class ASGIMessageQueue(Send):
         self._closed = True
         self._new_message_event.set()
 
-    async def __call__(self, message: Message):
-        """Send a message, putting it on the queue.
-
-        `RuntimeError` is raised if the queue has been closed using `.close()`.
-        """
+    async def _put_message(self, message: Message):
         if self._closed:
             raise RuntimeError("New messages cannot be sent after the queue is closed.")
 
         await self._message_queue.put(message)
         self._new_message_event.set()
+
+    async def __call__(self, message: Message):
+        """Send a message, putting it on the queue.
+
+        XXX: THREAD SAFE!!!
+
+        `RuntimeError` is raised if the queue has been closed using `.close()`.
+        """
+        coro = self._put_message(message)
+        if self._write_thread_safe:
+            await asyncio.wrap_future(
+                asyncio.run_coroutine_threadsafe(coro, self._loop)
+            )
+        else:
+            await coro
 
     def get_messages_nowait(self) -> List[Message]:
         """Returns all messages that are currently available (non-blocking).
@@ -194,9 +212,11 @@ class ASGIReceiveProxy:
 
     def __init__(
         self,
+        loop: asyncio.AbstractEventLoop,
         request_id: str,
         actor_handle: ActorHandle,
     ):
+        self._loop = loop
         self._queue = asyncio.Queue()
         self._request_id = request_id
         self._actor_handle = actor_handle
@@ -227,15 +247,25 @@ class ASGIReceiveProxy:
                 self._queue.put_nowait(e)
                 return
 
+    async def _get_next_message(self) -> Message:
+        if self._queue.empty() and self._disconnect_message is not None:
+            return self._disconnect_message
+
+        return await self._queue.get()
+
     async def __call__(self) -> Message:
         """Return the next message once available.
 
         This will repeatedly return a disconnect message once it's been received.
-        """
-        if self._queue.empty() and self._disconnect_message is not None:
-            return self._disconnect_message
 
-        message = await self._queue.get()
+        XXX: THREAD SAFE!!!
+        """
+        message = await asyncio.wrap_future(
+            asyncio.run_coroutine_threadsafe(
+                self._get_next_message(),
+                self._loop,
+            )
+        )
         if isinstance(message, Exception):
             raise message