--- a/python/ray/serve/_private/http_util.py
+++ b/python/ray/serve/_private/http_util.py
@@ -140,11 +140,13 @@ async def receive_http_body(scope, receive, send):
     return b"".join(body_buffer)
 
 
-class ASGIMessageQueue(Send):
+class MessageQueue(Send):
     """Queue enables polling for received or sent messages.
 
     This class assumes a single consumer of the queue (concurrent calls to
     `get_messages_nowait` and `wait_for_message` is undefined behavior).
+
+    This class implements the ASGI `Send` interface.
     """
 
     def __init__(self):