… python/ray/serve/multiplex.py
… python/ray/serve/tests/test_handle_streaming.py
… python/ray/serve/tests/test_http_util.py