@edoakes 2024-01-09
Streaming microbenchmark results show no significant difference.

Before (on master):
```
(ray) eoakes@Edwards-MacBook-Pro-2 serve % python _private/benchmarks/streaming_http_throughput.py
...
HTTP streaming throughput (num_replicas=1, tokens_per_request=1000, batch_size=10, use_intermediate_deployment=False): 242411.47 +- 4603.52 tokens/s
(ray) eoakes@Edwards-MacBook-Pro-2 serve % python _private/benchmarks/streaming_handle_throughput.py
...
DeploymentHandle streaming throughput (num_replicas=1, tokens_per_request=1000, batch_size=10): 16975.09 +- 1303.82 tokens/s
```

After (on this branch):
```
(ray) eoakes@Edwards-MacBook-Pro-2 serve % python _private/benchmarks/streaming_http_throughput.py
...
HTTP streaming throughput (num_replicas=1, tokens_per_request=1000, batch_size=10, use_intermediate_deployment=False): 243444.54 +- 5633.78 tokens/s
(ray) eoakes@Edwards-MacBook-Pro-2 serve % python _private/benchmarks/streaming_handle_throughput.py
...
DeploymentHandle streaming throughput (num_replicas=1, tokens_per_request=1000, batch_size=10): 16921.91 +- 599.67 tokens/s
```