@edoakes 2024-01-10
@alexeykudinkin to your question about the motivation:

I am splitting up the PRs so this one will have no behavioral change and should be a pure refactor, that's why we're still passing in the existing event loop rather than creating a standalone one. The following PR should be only a few additional lines of code.

There are multiple intertwined motivations for running the user code on its own loop:

- The primary reason is to productionize my prototype for late binding/new routing. In order for this to work, we need to ensure the call to get the queue length/start executing a request (1) cannot hang due to user code blocking the event loop and (2) run on the same thread as each other.
- More generally, I'd like to avoid having user code able to fully block our code in the replica (can help improve observability quite a lot) and make the concurrency model more explicit. Currently some of the methods implicitly run on separate threads/event loops (using `concurrency_group` support) which makes it very easy to have subtle race conditions. For example user health check methods are not currently thread safe because they're actually called on a separate event loop.
- We also currently rely on some very experimental/sketchy features of Ray to support our use case (actor method call stats dictionary & `concurrency_group` support in conjunction with `asyncio`). These aren't really used heavily elsewhere or well-tested, so I'd prefer to avoid them.