@alexeykudinkin 2024-01-10
@edoakes thanks for clarifying!

> More generally, I'd like to avoid having user code able to fully block our code in the replica (can help improve observability quite a lot) and make the concurrency model more explicit. Currently some of the methods implicitly run on separate threads/event loops (using concurrency_group support) which makes it very easy to have subtle race conditions. For example user health check methods are not currently thread safe because they're actually called on a separate event loop.

Am i reading this right that we want to get away from using concurrency-groups instead maintaining just a dichotomy of control-plane/user event-loops and running all the code in either of them?